agent/pyproject.toml
[project]
name = "spatialreal-agent"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.4",
    "livekit-plugins-spatialreal",
    "python-dotenv",
]

[tool.uv]
dev-dependencies = []
